Refactor: move to controller
